+++ R/utils_geometry.r
@@ -30,12 +30,17 @@
 #' Surface covered by a \code{LAS*} object. For \code{LAS} point clouds it is computed based on the
 #' convex hull of the points. For a \code{LAScatalog} it is computed as the sum of the bounding boxes
 #' of the files. For overlapping tiles the value may be larger than the total covered area because
-#' some regions are sampled twice.
+#' some regions are sampled twice. For a \code{LASheader} it is computed with the bounding box. The
+#' function \code{npoints} does what user may expect it to do and the function \code{density} is
+#' equivalent to \code{npoints(x)/area(x)}. As a consequence for the same file \code{area} applied on
+#' a LASheader or on a LAS can returns slighly different values.
 #'
 #' @param x An object of the class \code{LAS*}
 #' @param ... unused
 #'
-#' @return numeric. The area of the object computed in the same units as the coordinate reference system
+#' @return numeric. A number. Notice that for area the measure is in the same units as the coordinate
+#' reference system.
+#'
 #' @export
 #' @importMethodsFrom raster area
 #' @rdname area
@@ -46,20 +51,90 @@ setGeneric("area", function(x, ...)
 #' @rdname area
 setMethod("area", "LAS", function(x, ...)
 {
-  if (nrow(x@data) == 0)
-    return(0)
-
+  if (nrow(x@data) == 0) { return(0) }
   return(area_convex_hull(x@data$X, x@data$Y))
 })
+
+#' @export
+#' @rdname area
+setMethod("area", "LASheader", function(x, ...)
+{
+  PHB  <- x@PHB
+  area <- sum((PHB[["Max X"]] - PHB[["Min X"]]) * (PHB[["Max Y"]] - PHB[["Min Y"]]))
+  return(area)
+})
+
 
 #' @rdname area
 #' @export
 setMethod("area", "LAScatalog",  function(x, ...)
 {
   x <- x@data
-  area <- sum((x$Max.X - x$Min.X) * (x$Max.Y - x$Min.Y))
+  area <- sum((x[["Max.X"]] - x[["Min.X"]]) * (x[["Max.Y"]] - x[["Min.Y"]]))
   return(area)
 })
+
+# === npoints ====
+
+#' @rdname area
+#' @export
+setGeneric("npoints", function(x, ...)
+  standardGeneric("npoints"))
+
+#'
+#' @export
+#' @rdname area
+setMethod("npoints", "LAS", function(x, ...)
+{
+  return(nrow(x@data))
+})
+
+#' @export
+#' @rdname area
+setMethod("npoints", "LASheader", function(x, ...)
+{
+  return(x@PHB[["Number of point records"]])
+})
+
+
+#' @rdname area
+#' @export
+setMethod("npoints", "LAScatalog",  function(x, ...)
+{
+  return(sum(x[["Number.of.point.records"]]))
+})
+
+
+# === density ====
+
+#' @rdname area
+#' @export
+setGeneric("density", function(x, ...)
+  standardGeneric("density"))
+
+#' @export
+#' @rdname area
+setMethod("density", "LAS", function(x, ...)
+{
+  return(npoints(x)/area(x))
+})
+
+#' @export
+#' @rdname area
+setMethod("density", "LASheader", function(x, ...)
+{
+  return(npoints(x)/area(x))
+})
+
+#' @rdname area
+#' @export
+setMethod("density", "LAScatalog",  function(x, ...)
+{
+  return(npoints(x)/area(x))
+})
+
+# === internal ====
+
 
 area_convex_hull = function(x, y)
 {